fix(desktop): restore summarizer cost pill lost in work-surface refactor
The v0.1.16 3-column work-surface refactor dropped the summarizer
spend pill: the badge returned null on idle and the SlotPane gate
excluded idle status, so summarizer cost stopped surfacing.

Relocate SummarizerBadge beside its sole consumer in
SessionMetaFooter, make it self-contained (reads useSummarizerStatus
+ telemetry from props { sessionId }), and render the cost pill on
idle when at least one run has accrued (count > 0), avoiding a noisy
Sigma $0.0000 for never-summarized sessions. The summarizer-only Sigma
chip sits beside the agent-only SessionCostChip, preserving the
agent-vs-summarizer breakdown.
